pyproject.toml
[tool.poetry]
name = "dexy"
version = "0.1.0"
description = "An onchain AI agent"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.4"
langgraph = "^0.2.39"
coinbase-agentkit = "0.1.2"
coinbase-agentkit-langchain = "0.1.0"
langchain-community = "^0.3.18"
langchain-core = "^0.3.40"
pandas = "^2.2.3"
numpy = ">=1.26.4,<2"
matplotlib = "^3.10.1"
flask = "^3.1.0"

[tool.poetry.scripts]
start-agent = "chatbot:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
